Install and use the XSD schema file (closes #2).
We try the source tree location first, and fall back to the CMake
installation location if that does not exist. This will only work
if the API is installed in the location defined by CMake, but there
isn't really anything better we can do! And it should handle most
cases.

Also added a `parseFile` helper method (see issue #5) to test this
functionality.